Add CMake based builder for libpng
This is supposed to work for libpng 1.6.43, but not for libpng 1.6.34,
for instance.

Besides switching from MSBuild to CMake, we also no longer build zlib,
but instead rely on the zlib build already provided as dependency of
the respective PHP version.

To have the `.pdb`s, we work around the apparently common issue that
these are not properly installed.

We also append the soname suffix (`16` in this case) to the include
directory for BC reasons.