Ensure that we don't build against an unwanted crypto library (GH-17)
libzip can use several different crypto libraries, and checks their
availability by default.  It then processes this information in a
certain order, and uses the first of these libraries that are have been
found.

We always want to use the Windows cryptography library (bcrypt), so we
ensure that only support for this is enabled.