Add OpenLDAP MSYS2 builder
This does a full client build of OpenLDAP, using upstream's autotools
based build chain on MSYS2, but using MSVC build tools.  This requires
wrappers for the MSVC build tools; ar-lib and compile are taken
unmodified from automake; windres is a minimalist self-made wrapper.

We also need POSIX regex for the client tools (such as ldapsearch); we
use rxspencer[1], which we build on the fly and link in statically.

Using upstream's autotools based toolchain has the obvious drawback
that building now requires MSYS2 instead of "native" Windows tools
only.  On the other hand, instead of needing hand-made configuration
and Visual Studio solutions, what is volatile regarding updates of the
library, we only need a manageable patch set.  Plus, we get the client
tools basically for free.  If we want to distribute them is another
question.

[1] <https://github.com/garyhouston/rxspencer>