Add builder for libimagequant
This is supposed to work with the v2 based msvc branch, but not the
current rust based implementation.  As such, it should probably not be
used for production purposes, but seems to be good enough for some
experiments.

Since there are apparently some OMP constructs which are not supported
by MSVC, we disable these without further ado.