Fix #39: Static libavif looses debug info
We're resorting to drastic measures to avoid fiddling with the
internals of the libavif build system by applying `/Z7` instead of the
customary `/Zi`, so that debug info is embedded into the objects
instead of an external PDB.  While this increases the size of the
object files, the result when building the gd.dll should be the same.
And although this CMake option is only available as of version 3.25,
that shouldn't be an issue at least for the GH builds, since these ship
pretty recent CMake versions anyway.